redone work on redux storage
