+++ client/src/state/postsSlice.js
@@ -1,0 +1,27 @@
+import { createSlice } from "@reduxjs/toolkit";
+
+const initialState = {
+  posts: [],
+};
+
+export const postsSlice = createSlice({
+  name: "posts",
+  initialState,
+  reducers: {
+    setPosts: (state, action) => {
+      state.posts = action.payload.posts;
+    },
+    setPost: (state, action) => {
+      const updatedPosts = state.posts.map((post) => {
+        if (post._id === action.payload.post._id) return action.payload.post;
+        return post;
+      });
+      state.posts = updatedPosts;
+    },
+  },
+});
+
+// Action creators are generated for each case reducer function
+export const { setPosts, setPost } = postsSlice.actions;
+
+export default postsSlice.reducer;
